refact: move reporting functions to Report
