+++ scripts/build_ts.sh
@@ -1,4 +1,4 @@
 #!/usr/bin/env bash
 
-cp -r /home/yaro/projects/kulala.nvim/lua/kulala/parser/scripts/engines/javascript/lib/. /home/yaro/.cache/nvim/kulala/scripts/build/
-npm run build --prefix /home/yaro/.cache/nvim/kulala/scripts/build/
+cp -r lua/kulala/parser/scripts/engines/javascript/lib/. ${XDG_CACHE_HOME}/nvim/kulala/scripts/build/
+npm run build --prefix ${XDG_CACHE_HOME}/nvim/kulala/scripts/build/
